Revert "Refactor game subsystems into dedicated modules"
